requirements.txt
Django==1.11.12
django-crispy-forms==1.6.0
Pillow==5.1.0
django-tinymce==2.7.0
django-allauth==0.35.0
django-registration==2.1.2
certifi==2018.4.16
chardet==3.0.4
idna==2.6
oauthlib==2.0.7
python-openid==2.2.5
pytz==2018.4
requests==2.18.4
requests-oauthlib==0.8.0
urllib3==1.22
pkg-resources==0.0.0
